refactor(consume): simplify API behavior for direct URLs
Remove the --no-api-calls flag and instead make the behavior automatic:
- Direct GitHub release URLs: No API calls, no release page output
- Release specifiers (stable@latest): API calls required, includes release page

This provides better CI/UX without requiring additional flags.
